Conflicts:
	src/main/java/com/elster/jupiter/metering/rest/impl/MessageSeeds.java
	src/main/java/com/elster/jupiter/metering/rest/impl/MetrologyConfigurationResource.java
	src/main/java/com/elster/jupiter/metering/rest/impl/ResourceHelper.java

+++ coko/com.elster.jupiter.metering.rest/src/main/java/com/elster/jupiter/metering/rest/impl/MetrologyConfigurationResource.java
@@ -140,8 +140,12 @@ public class MetrologyConfigurationResource {
     public Response deleteMetrologyConfiguration(@PathParam("id") String id, MetrologyConfigurationInfo info) {
         MetrologyConfiguration metrologyConfiguration = resourceHelper.findAndLockMetrologyConfiguration(info);
 
+        if (metrologyConfiguration.isActive()) {
+            throw exceptionFactory.newException(MessageSeeds.YOU_CANNOT_REMOVE_ACTIVE_METROLOGY_CONFIGURATION);
+        }
+
         metrologyConfiguration.delete();
 
-        return Response.status(Response.Status.OK).entity(info).build();
+        return Response.status(Response.Status.OK).build();
     }
 }
